+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.0.2
 
 - Refactor: Improve handling of invalid tokens
+- Refactor: Improve handling of auth failed
 
 ## 0.0.1